requirements.txt
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.5
djangorestframework==3.11.0
drf-pdf==0.2.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
networkx==2.4
nltk==3.4.5
numpy==1.18.1
Pillow==7.0.0
# psycopg2==2.8.4
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
twilio==6.35.2
urllib3==1.25.7
whitenoise==5.0.1
xhtml2pdf==0.2.4
scipy==1.4.1
